Fix the overlaps counting
Fixed the overlaps counting, as the overlaps number is crazy after
legalization probably the algo does not work as intended.
